refactor(gui): modularize GUI components into separate files for better maintainability and readability
chore(gitignore): add .aider* to ignore list to prevent tracking of temporary files
