+++ src/gui/utilities.py
@@ -1,0 +1,5 @@
+from nicegui import ui
+from loguru import logger
+
+class Utilities:
+
